Remove RedrawEventsCleared + rename MainEventsCleared to AboutToWait
The idea that redraw events are dispatched with a specific ordering
that makes it possible to specifically report when we have finished
dispatching redraw events isn't portable and the way in which we
dispatched RedrawEventsCleared was inconsistent across backends.

More generally speaking, there is no inherent relationship between
redrawing and event loop iterations. An event loop may wake up at any
frequency depending on what sources of input events are being listened
to but redrawing is generally throttled and in some way synchronized
with the display frequency.

Similarly there's no inherent relationship between a single event loop
iteration and the dispatching of any specific kind of "main" event.

An event loop wakes up when there are events to read (e.g. input
events or responses from a display server / compositor) and goes back
to waiting when there's nothing else to read.

There isn't really a special kind of "main" event that is dispatched
in order with respect to other events.

For what MainEventsCleared is used for by applications then the most
portable interpretation of what applications want is to know when
the event loop is about to block and wait for new events.

In practice most applications almost certainly shouldn't care about
this because the frequency of event loop iterations is essentially
arbitrary but it's reasonably portable to expose this.

MainEventsCleared has been renamed to AboutToWait and is the last
event emitted in each iteration of the event loop.